feat : Impl Create Employee

+++ src/main/java/com/sprint/mission/sb8hrbankteamquerity/dto/employee/EmployeeCreateRequest.java
@@ -5,7 +5,7 @@ import java.time.Instant;
 public record EmployeeCreateRequest(
     String name,
     String email,
-    int departmentId,
+    Long departmentId,
     String position,
     Instant hireDate,
     String memo
